# contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: DriveRAG API
  version: "0.2.0"
  description: |
    Agentic Knowledge Graph RAG system API.
    This is the contract that the backend implements and the frontend consumes.
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.dingdong-rag.com/api/v1
    description: Production
paths:
...
  /knowledge/entities:
    post:
      operationId: createEntity
      summary: Create a knowledge entity
      tags: [Knowledge]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntityCreate"
      responses:
        "201":
          description: Entity created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntityResponse"
  /knowledge/entities/{entity_id}:
    get:
      operationId: getEntity
      summary: Get entity with relationships
      tags: [Knowledge]
      parameters:
        - name: entity_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Entity with relationships
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntityWithRelationships"
        "404":
          description: Entity not found
  /knowledge/relationships:
    post:
      operationId: createRelationship
      summary: Create a relationship between entities
      tags: [Knowledge]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RelationshipCreate"
      responses:
        "201":
          description: Relationship created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RelationshipResponse"
        "404":
          description: Entity not found
  /knowledge/search/vector:
    post:
      operationId: vectorSearch
      summary: Search by vector similarity
      tags: [Knowledge]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VectorSearchRequest"
      responses:
        "200":
          description: Vector search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VectorSearchResult"
  /knowledge/search/hybrid:
    post:
      operationId: hybridSearch
      summary: Hybrid search combining vector and graph
      tags: [Knowledge]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HybridSearchRequest"
      responses:
        "200":
          description: Hybrid search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HybridSearchResult"
  /knowledge/graph:
    get:
      operationId: getGraph
      summary: Get graph visualization data
      tags: [Knowledge]
      parameters:
        - name: document_id
          in: query
          schema:
            type: string
            format: uuid
        - name: entity_types
          in: query
          description: Comma-separated entity types
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        "200":
          description: Graph visualization data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphVisualization"
...
  /settings/models/available:
    get:
      operationId: getAvailableModels
      summary: Get chat models available to the current user
      description: |
        Returns chat models filtered by which providers have API keys,
        combining server-level env vars with the user's per-user keys.
      tags: [Settings]
      responses:
        "200":
          description: Available chat models
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailableModelsResponse"
        "401":
          description: Not authenticated
...
  /settings/models:
    get:
      operationId: getModelSettings
      summary: Get current user's model settings
      description: Returns the user's selected models and API key presence flags (keys never returned in plaintext).
      tags: [Settings]
      responses:
        "200":
          description: Model settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelSettingsResponse"
        "401":
          description: Not authenticated
    put:
      operationId: updateModelSettings
      summary: Update current user's model settings
      description: |
        Upserts model selections and API keys for the authenticated user.
        API key handling:
        - Empty string `""` — leave the existing key unchanged
        - `null` — clear (delete) the stored key
        - Any other string — encrypt and store as new key
      tags: [Settings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateModelSettingsRequest"
      responses:
        "200":
          description: Updated model settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelSettingsResponse"
        "401":
          description: Not authenticated
components:
  schemas:
...
    EntityCreate:
      type: object
      required: [entity_type, name]
      properties:
        entity_type:
          type: string
        name:
          type: string
          maxLength: 500
        description:
          type: string
          nullable: true
        properties:
          type: object
          nullable: true
        source_document_id:
          type: string
          format: uuid
          nullable: true
    EntityResponse:
      type: object
      required: [id, neo4j_id, entity_type, name, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        neo4j_id:
          type: string
        entity_type:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        properties:
          type: object
          nullable: true
        source_document_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        relationship_count:
          type: integer
          default: 0
    EntityWithRelationships:
      allOf:
        - $ref: "#/components/schemas/EntityResponse"
        - type: object
          properties:
            relationships:
              type: array
              items:
                $ref: "#/components/schemas/RelationshipResponse"
    RelationshipCreate:
      type: object
      required: [relationship_type, source_entity_id, target_entity_id]
      properties:
        relationship_type:
          type: string
        source_entity_id:
          type: string
          format: uuid
        target_entity_id:
          type: string
          format: uuid
        properties:
          type: object
          nullable: true
        weight:
          type: number
          minimum: 0
          maximum: 1
          default: 1.0
        source_document_id:
          type: string
          format: uuid
          nullable: true
    RelationshipResponse:
      type: object
      required: [id, neo4j_id, relationship_type, source_entity_id, target_entity_id, created_at]
      properties:
        id:
          type: string
          format: uuid
        neo4j_id:
          type: string
        relationship_type:
          type: string
        source_entity_id:
          type: string
          format: uuid
        target_entity_id:
          type: string
          format: uuid
        source_entity_name:
          type: string
          nullable: true
        target_entity_name:
          type: string
          nullable: true
        properties:
          type: object
          nullable: true
        weight:
          type: number
        source_document_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
    VectorSearchRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 2000
        top_k:
          type: integer
          minimum: 1
          maximum: 100
          default: 10
        document_id:
          type: string
          format: uuid
          nullable: true
        similarity_threshold:
          type: number
          minimum: 0
          maximum: 1
          default: 0.7
    VectorSearchResult:
      type: object
      required: [chunk_id, document_id, content, similarity]
      properties:
        chunk_id:
          type: string
          format: uuid
        document_id:
          type: string
          format: uuid
        content:
          type: string
        similarity:
          type: number
        metadata:
          type: object
          nullable: true
    HybridSearchRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 2000
        top_k:
          type: integer
          minimum: 1
          maximum: 100
          default: 10
        vector_weight:
          type: number
          minimum: 0
          maximum: 1
          default: 0.5
        document_id:
          type: string
          format: uuid
          nullable: true
    HybridSearchResult:
      type: object
      required: [content, source, score]
      properties:
        content:
          type: string
        source:
          type: string
          enum: [vector, graph, both]
        score:
          type: number
        document_id:
          type: string
          format: uuid
          nullable: true
        chunk_id:
          type: string
          format: uuid
          nullable: true
        entity_id:
          type: string
          format: uuid
          nullable: true
        metadata:
          type: object
          nullable: true
    GraphNode:
      type: object
      required: [id, label, type]
      properties:
        id:
          type: string
        label:
          type: string
        type:
          type: string
        description:
          type: string
          nullable: true
        properties:
          type: object
          nullable: true
        size:
          type: number
          default: 1.0
    GraphEdge:
      type: object
      required: [source, target, label]
      properties:
        source:
          type: string
        target:
          type: string
        label:
          type: string
        weight:
          type: number
          default: 1.0
    GraphVisualization:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/GraphNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/GraphEdge"
        total_nodes:
          type: integer
        total_edges:
          type: integer
...
    ApiKeyStatus:
      type: object
      required: [has_key]
      properties:
        has_key:
          type: boolean
          description: True if an API key is stored for this provider.
    ModelSettingsResponse:
      type: object
      required: [chat_model, ingestion_model, embedding_model, anthropic_api_key, openai_api_key, dashscope_api_key, openrouter_api_key]
      properties:
        chat_model:
          type: string
          example: dashscope/qwen3-max
        ingestion_model:
          type: string
          example: dashscope/qwen3-max
        embedding_model:
          type: string
          example: openai/text-embedding-3-small
        anthropic_api_key:
          $ref: "#/components/schemas/ApiKeyStatus"
        openai_api_key:
          $ref: "#/components/schemas/ApiKeyStatus"
        dashscope_api_key:
          $ref: "#/components/schemas/ApiKeyStatus"
        openrouter_api_key:
          $ref: "#/components/schemas/ApiKeyStatus"
    UpdateModelSettingsRequest:
      type: object
      properties:
        chat_model:
          type: string
          nullable: true
        ingestion_model:
          type: string
          nullable: true
        embedding_model:
          type: string
          nullable: true
        anthropic_api_key:
          type: string
          nullable: true
          description: |
            "" = leave unchanged, null = clear the stored key, any other value = new key to store
        openai_api_key:
          type: string
          nullable: true
          description: |
            "" = leave unchanged, null = clear the stored key, any other value = new key to store
        dashscope_api_key:
          type: string
          nullable: true
          description: |
            "" = leave unchanged, null = clear the stored key, any other value = new key to store
        openrouter_api_key:
          type: string
          nullable: true
          description: |
            "" = leave unchanged, null = clear the stored key, any other value = new key to store
    ModelOption:
      type: object
      required: [provider, model_id, label]
      properties:
        provider:
          type: string
          example: Anthropic
        model_id:
          type: string
          example: anthropic/claude-sonnet-4-20250514
        label:
          type: string
          example: Claude Sonnet 4
    AvailableModelsResponse:
      type: object
      required: [models]
      properties:
        models:
          type: array
          items:
            $ref: "#/components/schemas/ModelOption"
          description: Chat models filtered by providers with available API keys.
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []
...